centralize all state logic and add replay feature
