+++ client/src/components/WordGrid.jsx
@@ -1,15 +1,15 @@
 import WordGridRow from "./WordGridRow";
 
-function WordGrid({ word, currentRow }) {
-    
+function WordGrid() {
+
     return (
         <div className="word-grid">
-            <WordGridRow currentRow={currentRow} word={word} row={0}/>
-            <WordGridRow currentRow={currentRow} word={word} row={1}/>
-            <WordGridRow currentRow={currentRow} word={word} row={2}/>
-            <WordGridRow currentRow={currentRow} word={word} row={3}/>
-            <WordGridRow currentRow={currentRow} word={word} row={4}/>
-            <WordGridRow currentRow={currentRow} word={word} row={5}/>
+            <WordGridRow row={0}/>
+            <WordGridRow row={1}/>
+            <WordGridRow row={2}/>
+            <WordGridRow row={3}/>
+            <WordGridRow row={4}/>
+            <WordGridRow row={5}/>
         </div>
 
     );
