+++ server/events/gameEvents.js
@@ -1,4 +1,5 @@
 const {isValidWord} = require('../utils/words');
+const { addUser, getRoomUserList, removeUser } = require("../utils/users");
 
 const letterReceiveHandler = (io, socket) => {
     socket.on("letter_input", ({ letter, room }) => {
